Add information on filters, as well as checks
Previously, there was no information in the README about filters,
and running Serrano::Filters in a repl didn't provide any information
(although we could read the code in the file).

This commit adds information about filters to the README, provides
access to filter names and information in the library, and checks
before running a request to make sure that filters are formatted as
a hash and that the filters provided are valid filters.